+++ src/index.ts
@@ -21,6 +21,8 @@ export * from './Divider';
 export * from './Dropdown';
 export * from './Empty';
 export * from './Form';
+export * from './Grid';
+export * from './Image';
 export * from './Input';
 export * from './List';
 export * from './Menu';